requirements.txt
flask==2.3.3
pandas==2.1.0
python-dotenv==1.0.0
openai==1.12.0
groq==0.4.2
boto3==1.34.34
openpyxl==3.1.2
xlrd==2.0.1
PyPDF2==3.0.1
requests==2.31.0
anthropic==0.18.1
google-generativeai==0.3.2
